fixed lints
